Obligatorio-SQL/src/server-flask/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
mysql-connector-python==8.2.0
python-dotenv==1.0.0

